% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{episode_list}
\alias{episode_list}
\title{Basic information about episodes}
\format{A data frame with 287 rows and 7 variables.}
\source{
\url{https://southpark.wikia.com/wiki/Portal:Scripts}
}
\usage{
episode_list
}
\description{
A dataset containing information about South Park episodes,
  seasons they belong to and links to their scripts. It can
  be recreated using \code{\link{fetch_episode_list}()}.
}
\details{
\describe{
  \item{season_episode_number}{Number of an episode in a season.}
  \item{season_link}{
    URL link to a page of season scripts. That page contains
    a list of episodes and links to their transcripts.
    Example: \url{http://southpark.wikia.com/wiki/Portal:Scripts/Season_One"}}
  \item{episode_link}{
    URL link to an episode transcript.
    Example: \url{http://southpark.wikia.com/wiki/Cartman_Gets_an_Anal_Probe/Script}}
  \item{episode_name}{Official name of an episode.}
  \item{season_name}{String in a format: Season ___ (number as a word).}
  \item{season_number}{Numeric number of a season.}
  \item{season_year}{A year when an episode was released.}
}
}
\keyword{datasets}
