% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southpark_crawler.R
\name{fetch_episode}
\alias{fetch_episode}
\title{Fetch transcript for a single episode.}
\usage{
fetch_episode(episode_link)
}
\arguments{
\item{episode_link}{}
}
\description{
Fetch transcript for a single episode.
}
