% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{fetch_episode_list}
\alias{fetch_episode_list}
\title{Fetch list of all episodes.}
\usage{
fetch_episode_list()
}
\value{
Data frame with list of episodes.
}
\seealso{
\code{\link{episode_list}} dataset.
}
