% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb_ratings.R
\name{fetch_ratings_old}
\alias{fetch_ratings_old}
\title{Fetch ratings for all selected seasons.}
\usage{
fetch_ratings_old(season_numbers)
}
\arguments{
\item{season_numbers}{Numeric season number.}
}
\description{
**DO NOT USE** - it violates IMDB conditions of use. It does
  not allow web scraping. It provides datasets that can
  be parsed accordingly.
}
\seealso{
\code{\link{fetch_ratings}()} - use this function insted.
}
