% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{fetch_season_episode_list}
\alias{fetch_season_episode_list}
\title{Fetch episode list for a season}
\usage{
fetch_season_episode_list(season_link)
}
\arguments{
\item{season_link}{Link to a page of season scripts.}
}
\value{
Data frame with list of episodes.
}
\seealso{
\code{\link{episode_list}} dataset.
}
