% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southpark_ratings.R
\name{fetch_season_ratings}
\alias{fetch_season_ratings}
\title{Fetch IMDB ratings for one season.}
\usage{
fetch_season_ratings(season_number)
}
\arguments{
\item{season_number}{}
}
\description{
Fetch IMDB ratings for one season.
}
