% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb_ratings.R
\name{fetch_season_ratings_old}
\alias{fetch_season_ratings_old}
\title{Fetch IMDB ratings for one season.}
\usage{
fetch_season_ratings_old(season_number)
}
\description{
Fetch IMDB ratings for one season.
}
