% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_character_sentiment}
\alias{plot_character_sentiment}
\title{Plots sentiment over episodes for selected characters.}
\usage{
plot_character_sentiment(words, characters)
}
\arguments{
\item{characters}{}
}
\description{
Plots sentiment over episodes for selected characters.
}
