% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{preprocess_characters}
\alias{preprocess_characters}
\title{Preprocess character names.}
\usage{
preprocess_characters(characters)
}
\arguments{
\item{characters}{Character vector.}
}
\value{
Modified character vector.
}
\description{
This function strips semicolons and double
  quotes from character names. It also converts
  all characters to lower case.
}
