% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Preprocess text to keep only alphanum character, spaces and apostrophes.}
\usage{
preprocess_text(text)
}
\arguments{
\item{text}{Character vector of text lines.}
}
\value{
Modified character vector.
}
\description{
Preprocess text to keep only alphanum character, spaces and apostrophes.
}
