% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southparkr.R
\docType{package}
\name{southparkr-package}
\alias{southparkr}
\alias{southparkr-package}
\title{southparkr: South Park script crawler}
\description{
southparkr: South Park script crawler
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pdrhlik/southparkr}
  \item Report bugs at \url{https://github.com/pdrhlik/southparkr/issues}
}

}
\author{
\strong{Maintainer}: Patrik Drhlik \email{patrik.drhlik@gmail.com}

}
