% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{swear_words}
\alias{swear_words}
\title{List of words that are considered naughty.}
\usage{
swear_words()
}
\value{
Character vector of swear words.
}
\description{
This function will be replaced in the future
  by the \href{https://github.com/pdrhlik/sweary}{sweary}
  package.
}
