% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{top_n_characters}
\alias{top_n_characters}
\title{Select top N talking characters.}
\usage{
top_n_characters(words, n)
}
\arguments{
\item{words}{A data frame of words returned by
\code{\link{process_episode_words}}.}

\item{n}{Numeric value - how many TOP talking
characters to select.}
}
\value{
Character vector of character names.
}
\description{
Select top N talking characters.
}
