% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{compare_two_characters}
\alias{compare_two_characters}
\title{Compare two characters and decide which one is naughtier.}
\usage{
compare_two_characters(char2, char1, words)
}
\arguments{
\item{words}{}
}
\description{
Compare two characters and decide which one is naughtier.
}
