% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{fetch_all_episodes}
\alias{fetch_all_episodes}
\title{Fetch scripts for all episodes.}
\usage{
fetch_all_episodes(episode_list)
}
\arguments{
\item{episode_list}{\code{\link{episode_list}} dataset.}
}
\value{
A data frame with episode lines.
}
\seealso{
\code{\link{episode_lines}} dataset.
}
