% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southpark_crawler.R
\name{fetch_season_episode_list}
\alias{fetch_season_episode_list}
\title{Fetch episode list for a season data_frame.}
\usage{
fetch_season_episode_list(season_link, season_name, season_number, season_year)
}
\arguments{
\item{season_year}{}
}
\value{
data_frame
}
\description{
Fetch episode list for a season data_frame.
}
