% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb_ratings.R
\name{fetch_season_ratings_old}
\alias{fetch_season_ratings_old}
\title{Fetch IMDB ratings for one season.}
\usage{
fetch_season_ratings_old(season_number)
}
\arguments{
\item{season_number}{Numeric season number.}
}
\description{
**DO NOT USE** - it violates IMDB conditions of use. It does
  not allow web scraping. It provides datasets that can
  be parsed accordingly.
}
\seealso{
\code{\link{fetch_ratings}()} - use this function insted.
}
