% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{plot_swear_word_comparison}
\alias{plot_swear_word_comparison}
\title{Plot comparison of naughty words proportion among selected characters.}
\usage{
plot_swear_word_comparison(main_character, other_characters, words,
  text_size = 30)
}
\arguments{
\item{other_characters}{}
}
\description{
Plot comparison of naughty words proportion among selected characters.
}
