% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southpark_crawler.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Preprocess text to keep only alphanum character, spaces and apostrophes.}
\usage{
preprocess_text(text)
}
\description{
Preprocess text to keep only alphanum character, spaces and apostrophes.
}
