% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{process_episode_words}
\alias{process_episode_words}
\title{Processes lines and ratings and creates a data frame usable for analyses.}
\usage{
process_episode_words(lines, ratings)
}
\arguments{
\item{ratings}{}
}
\description{
Processes lines and ratings and creates a data frame usable for analyses.
}
